requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0

# CLI and UI
rich>=13.0.0
click>=8.1.0

# Data processing
numpy>=1.24.0
openpyxl>=3.1.0

# Web scraping
selenium>=4.15.0
lxml>=4.9.0

# HTTP and networking
urllib3>=2.0.0
httpx>=0.25.0

# Report generation
jinja2>=3.1.0
weasyprint>=60.0
reportlab>=4.0.0

# Email and contact validation
email-validator>=2.1.0
phonenumbers>=8.13.0

# Utilities
pydantic>=2.5.0
tqdm>=4.66.0
colorama>=0.4.6

# Optional: Machine Learning (for advanced features)
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
